+++ bitcount/bitcnt_4.c
@@ -35,7 +35,7 @@ static char bits[256] =
 **        omitted.
 */
 
-int CDECL ntbl_bitcnt(long x)
+int CDECL ntbl_bitcnt(uint32_t x)
 {
       int cnt = bits[(int)(x & 0x0000000FL)];
 
@@ -49,7 +49,7 @@ int CDECL ntbl_bitcnt(long x)
 **  Count bits in each byte
 */
 
-int CDECL btbl_bitcnt(long x)
+int CDECL btbl_bitcnt(uint32_t x)
 {
       int cnt = bits[ ((char *)&x)[0] & 0xFF ];
 
@@ -65,7 +65,7 @@ int CDECL btbl_bitcnt(long x)
 
 main(int argc, char *argv[])
 {
-      long n;
+      uint32_t n;
 
       while(--argc)
       {
